backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
scikit-learn==1.5.2
numpy==2.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
